src/lambdas/split_n_convert_pdf/requirements.txt
boto3
PyMuPDF
pyyaml
loguru
pillow